create assets directory with national and state asset splits
