+++ cv-assets/cv_assets/utils/__init__.py
@@ -1,1 +1,2 @@
+from cv_assets.utils.load_pg_table import load_table_from_parquet
 from cv_assets.utils.shell_cmd import run_shell_cmd
